+++ src/components/viewport-elements/game-context-zones/game-context-zones.js
@@ -1,5 +1,0 @@
-import { GameContextZones } from "./GameContextZones.js";
-
-customElements.define("game-context-zones", GameContextZones);
-
-export * from "./GameContextZones.js";