+++ src/components/viewport-elements/game-theme-zones/game-theme-zones.js
@@ -1,5 +1,0 @@
-import { GameThemeZones } from "./GameThemeZones.js";
-
-customElements.define("game-theme-zones", GameThemeZones);
-
-export * from "./GameThemeZones.js";