+++ src/components/viewport-elements/game-zone-indicator/game-zone-indicator.js
@@ -1,0 +1,5 @@
+import { GameZoneIndicator } from "./GameZoneIndicator.js";
+
+customElements.define("game-zone-indicator", GameZoneIndicator);
+
+export * from "./GameZoneIndicator.js";